backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
fpdf2==2.7.6
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
sendgrid==6.11.0
Pillow==10.1.0
# Email alternatives
email-validator==2.1.0